Holy crap, huge perf improvement using the minimaxer lib properly/
